feat: noop plugin for gateway load testing
This adds an example plugin that can be useful for benchmarking
the gateway frontend without adding any storage I/O on the
backend. This can also be a useful template for developers
looking to implement a new plugin for the gateway.

Also add a build verification to the go workflow for every plugin
in the plugins directory.