fix: Azurite docker image server certificate configuration
The azurite container required HTTPS access. The certs were
generated with: mkcert localhost 127.0.0.1 ::1 azurite
The azurite seems to be the important name.

The certs are also copied into the system trust for the gateway
as the sdk client needs to be able to trust the certs from the
azurite server.

This also bumps the go module to 1.21 for the error:
/app/s3api/utils/sign_hack.go:43:67: reflect.ValueOf((*ignoredHeaders)[0]).FieldByName("Rule").Elem().Clear undefined (type reflect.Value has no field or method Clear)
And uses latest golang for Docker.dev.