fix: v4auth uri escape paths in canonical request
According to the docs:
https://docs.aws.amazon.com/IAM/latest/UserGuide/create-signed-request.html
The Canonical URIL and Canonical Query String need to be passed
through UriEncode() when generating the string to sign.

Its likely that the signer we are using is expecting this to
already be escaped under normal use. But for our purposes
we still need this to be escaped before generating the signature
to compare.

Fixes #781